Refactor : DTO를 도메인 계층으로 이동
