+++ src/main/java/com/eduquest/backend/infrastructure/s3/client/EduQuestS3Client.java
@@ -2,7 +2,7 @@ package com.eduquest.backend.infrastructure.s3.client;
 
 import com.eduquest.backend.common.exception.EduQuestException;
 import com.eduquest.backend.domain.file.component.CustomS3Client;
-import com.eduquest.backend.infrastructure.s3.dto.S3FileDto;
+import com.eduquest.backend.domain.file.dto.S3FileDto;
 import com.eduquest.backend.infrastructure.s3.exception.S3ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
